update fake eval, lint, fix message delete condition with badtraslate controller

+++ src/structures/BTChannelController.ts
@@ -10,14 +10,14 @@ export default class BTChannelController {
   private ratelimitCache: BaseCollection<string, number>;
   private _assyst: Assyst;
 
-  constructor(assyst: Assyst, channels?: string[]) {
+  constructor (assyst: Assyst, channels?: string[]) {
     this._assyst = assyst;
     this.channels = channels || [];
     this.webhookCache = new BaseCollection({ limit: 100 });
     this.ratelimitCache = new BaseCollection({ limit: 100 });
   }
 
-  public async init(): Promise<void> {
+  public async init (): Promise<void> {
     for (const channel of this.channels) {
       const webhooks: BaseCollection<string, Webhook> = await this._assyst.rest.fetchChannelWebhooks(channel);
       for (const webhook of webhooks.values()) {
@@ -26,12 +26,12 @@ export default class BTChannelController {
     }
 
     this._assyst.client.on('messageCreate', async ({ message }) => {
-      if (!this.channels.includes(message.channelId) || message.author.bot) return;
+      if (!this.channels.includes(message.channelId) || message.author.bot || !message.content) return;
 
       const userRatelimit = this.ratelimitCache.get(message.author.id);
       if (userRatelimit && Date.now() - userRatelimit < Ratelimit) {
         const timeLeft = (userRatelimit + Ratelimit) - Date.now();
-        await message.delete();
+        if (!message.deleted) await message.delete();
         const response = await message.reply(`${message.author.mention} try again in ${(timeLeft / 1000).toFixed(2)} seconds!`);
         setTimeout(() => response.delete(), 1500);
         return;
@@ -42,7 +42,7 @@ export default class BTChannelController {
     });
   }
 
-  public async handle(message: Message, attempt?: number) {
+  public async handle (message: Message, attempt?: number) {
     // prevent infinite recursion
     if (attempt && attempt > 5) {
       return;
@@ -50,18 +50,18 @@ export default class BTChannelController {
     const translatedText = await this.translate(message.content.substr(0, 1000), 3);
     const webhook = await this.getWebhookOrCreate(message.author.username, message.channelId);
     try {
-      await this.executeWebhook(webhook!, translatedText, message.author.avatarUrlFormat("png"));
-    } catch(e) {
+      await this.executeWebhook(webhook!, translatedText, message.author.avatarUrlFormat('png'));
+    } catch (e) {
       // Webhook no longer exists?
       this.webhookCache.delete(webhook!.id);
       // Retry
       await this.handle(message, (attempt || 0) + 1);
     }
 
-    await message.delete();
+    if (!message.deleted) await message.delete();
   }
 
-  public async executeWebhook(webhook: Webhook, content: string, avatarUrl: string): Promise<void> {
+  public async executeWebhook (webhook: Webhook, content: string, avatarUrl: string): Promise<void> {
     await webhook.execute({
       content: content.substr(0, 1999),
       avatarUrl,
@@ -71,23 +71,23 @@ export default class BTChannelController {
     });
   }
 
-  public async translate(text: string, hops?: number): Promise<string> {
+  public async translate (text: string, hops?: number): Promise<string> {
     // ONLY translate first 1000 chars
     // Yandex doesn't like >1k character translations (takes very long!)
     return this._assyst.customRest.translate(text.substr(0, 1000), hops || 6).then(r => r.text);
   }
 
-  public async getWebhookOrCreate(username: string, channelId: string): Promise<Webhook | undefined> {
+  public async getWebhookOrCreate (username: string, channelId: string): Promise<Webhook | undefined> {
     let webhook = this.webhookCache.find(w => w.name === username && w.channelId === channelId);
     if (!webhook) {
       try {
         webhook = await this._assyst.client.rest.createWebhook(channelId, {
           name: username
         });
-      } catch(e) {
+      } catch (e) {
         const lru = this.webhookCache.first();
         if (lru) {
-          await this._assyst.rest.deleteWebhook(lru.id)
+          await this._assyst.rest.deleteWebhook(lru.id);
           this.webhookCache.delete(lru.id);
           webhook = await this._assyst.client.rest.createWebhook(channelId, {
             name: username
